Add `cargo clippy` to CI (#2067)
<!-- ELLIPSIS_HIDDEN -->


> [!IMPORTANT]
> Add `cargo clippy` to CI and fix clippy warnings across the codebase
to enforce Rust linting rules.
> 
>   - **CI Integration**:
> - Add `cargo clippy` to `.github/workflows/primary.yml` for Rust
linting.
>     - Introduce steps for standard and WASM targets.
>   - **Code Changes**:
> - Fix clippy warnings by removing unnecessary references in
`attribute.rs`, `expression.rs`, and `mod.rs`.
> - Update error handling in `to_baml_arg.rs` to use `ArgCoerceError`
instead of `()`.
> - Simplify error handling in `repr.rs` by removing redundant
`or_else`.
>     - Correct scope usage in `expr_fns.rs` to avoid mutable borrow.
>     - Adjust logging and serialization in `logger.rs` and `lib.rs`.
>     - Update `Cargo.toml` files to allow specific clippy lints.
>   - **Miscellaneous**:
>     - Fix typos and improve code readability in various files.
> - Update test expectations in `input-output.test.ts` for consistency.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis"
src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=BoundaryML%2Fbaml&utm_source=github&utm_medium=referral)<sup>
for 91dd423355b8d1544f93c2881d9ba45f12c16a47. You can
[customize](https://app.ellipsis.dev/BoundaryML/settings/summaries) this
summary. It will automatically update as commits are pushed.</sup>


<!-- ELLIPSIS_HIDDEN -->

+++ engine/baml-lib/baml-core/src/validate/validation_pipeline/validations/template_strings.rs
@@ -91,11 +91,7 @@ pub(super) fn validate(ctx: &mut Context<'_>) {
                     let start_offset = pspan.start + range.start;
                     let end_offset = pspan.start + range.end;
 
-                    let span = Span::new(
-                        pspan.file.clone(),
-                        start_offset as usize,
-                        end_offset as usize,
-                    );
+                    let span = Span::new(pspan.file.clone(), start_offset, end_offset);
 
                     ctx.push_error(DatamodelError::new_validation_error(
                         &format!("Error parsing jinja template: {}", e),
