+++ engine/baml-lib/baml-core/src/validate/validation_pipeline/validations/tests.rs
@@ -63,11 +63,7 @@ pub(super) fn validate(ctx: &mut Context<'_>) {
                         let start_offset = expr_span.start + range.start;
                         let end_offset = expr_span.start + range.end;
 
-                        let span = Span::new(
-                            expr_span.file.clone(),
-                            start_offset as usize,
-                            end_offset as usize,
-                        );
+                        let span = Span::new(expr_span.file.clone(), start_offset, end_offset);
 
                         ctx.push_error(DatamodelError::new_validation_error(
                             &format!("Error parsing jinja template: {}", e),
